type: feat
adding the python dependency setup from the course repository. adding
some minimal code from the navigation notebook to a python script so I
can verify that I'm able to bring up the unity environment and get it to
print out some basic stats about the environment. started filling in
some dependencies on another script that is a starting place for an
environment runner.

addresses: having a working environment to train my task in.