type: chore and docs
Now that I've ensured the environment would work properly, I wanted to
find a more seamless way to install the dependencies to make it simpler
for the project evaluators. I found github issue because pypi had
stopped supporting the version of pytorch that this project used, so I
chose to download a local wheel file from pytorch for the version needed
and install it through requirements.txt. I also didn't end up needing
setup.py after this so I removed that too. The build directory had many
things in it but when I deleted themt he environment still worked so I
just trimmed the overhead there  until I find I need it.

addresses: having a project that is easy to run.