type: feature, chore
implementing the datastructure for pooling experience for replay, breaking up the main function into several sub modules of related functions. Fun with generators! turns out they are a great way to represent several operations on simple state over a series of iterations as function calls instead of having lots of loops everywhere, with the added benefit of potentially being infinite.