type: chore
ok, so now that I'm trying to run it I'm getting all the fun errors. So for one, windows didn't have the dlls for cuda so I had to figuer that out and add it to the readme, then a lot of things were just plain wrong with my imports and naming convention inconsistencies, so changed a lot of those. Still not running bu tI'm satisfied that every change represented here was a step in the right direction, so going to commit what I have.