type: chore
one by one fixing things that weren't working when running the code. things like using higher order function in the environment method so I could pass in reset or step. fixing the datatype used for the tensor extraction of actions by changing it to 'long' from float. lambdas that don't take an argument need *args not _, fixing parameter passing for a number of modules, and function composition to be clearer, added names for learning purposes.