type experiment
changing the update every hyperparameter to be every 10 steps instead of 4. This affects both how often teh agent learns and how often the soft update occurs. It means more actions are taken in the environment before each learning interval. It gives more opportunities for the agent to accumulate a negative score before the model tries to optimize.