type: feature
I added a timer counter. The udacity project for DQN had the agent learning on a set schedule that was independent of whether or not the episode that the agent was going through had finished or not, it was just a constant cadence. I'm not sure if that will improve things over mine which reset its learning according to a cadence that reset every episode. after running 500 episodes I didn't see much change, but I'll leave it in there to be consistent for now.