type: experiment
increasing number of episodes to 2000. I did see that the scores got higher than they had before but still oscillating quite a bit. Going to try another with even more episodes and also increasing the buffer size for the experience dataset to something much higher, now realizing that forgetting the old experiences might be making it harder to learn once we are on a much higher episode.