type: chore
I changed where I am converting the type of the actions tensor in the experiences to when they are sampled instead of  right as they are gathered. I didn't think this should make a big difference because it is essentially moving a conversion from when the value is used to when it is first retrieved and there isn't anything done to it in the meantime. But that is the way it is in the udacity DQN code and at this point I'm just trying to achieve parity with that.