type: experiment
attempting to initialize my weights using the xavier uniform random distribution as recommended by someone on the forums, owing to the whitepaper by Yoshua Bengio and Xavier Glorot. The hope was that with this normalized initialization I in case I had a vanishing or exploding gradient occurring because of the deeper network that the weight initializations would have counteracted that. But it looks like no dice, there is another problem preventing convergence that might be due to the correlation between learning and target network, not sure.