type experiment
putting the learning rate, tau and the update cadence back to the values from udacity's project, leaving batch size large as the only change. This is because I noticed one other difference from the DQN code and mine and that is unity was not accepting my numpy int64 type actions, and so I had casst them to int32s before returning them from the step function, when it turns out I could have just done it right in time to send it to the unity enviornment and left them as numpy int64 values the rest of the time. this doesn't fix my issue but it made the scores higher on average I think once again, like it is finding a better local minima to get stuck in.  now maybe if I combine this with more experience memory or a different value for tau or some other tweak to the hyperparameters.