type chore
couple things here, first, added  tests for epsilon greedy that exposed the bug that's probably been preventing me from learning, my reinforcement learning alogrithm's epsilon greedy policy strategy code was wrong. I was trying to do something fancy with short circuiting boolean condtional expressions, and that just doesn't work because python doesn't return the results it evaluates the whole thing and  unfortunately a falsy index fromt he action values tensor meant use the random choice, could be even worse I was potentially always either doing a random thing or hardcoding zero. That would explain lack of ability to learn. Additionally I updated the readme for a few things and added a printout of my episodes chart to show that the agent was in fact learning.

After fixing this bug, the agent was able to learn to get a score of 13 in only 609 episodes.

while I was on a roll, went ahead and did the first draft of the code that will load and run the agent when it is fully trained.